refactor(Validator.java): 클래스 분리 리팩토링
숫자야구게임 기능 중 유효성 체크하는 메소드를 별도 클래스로 분리한다.
- 숫자야구게임 관련 유효성 체크 메소드들을 정리한 Validator.java 생성
- 실행/게임 진행/유효성 체크 3가지 목적에 맞게 클래스 분리
