+++ src/main/java/baseballgame/BaseBallGame.java
@@ -2,30 +2,38 @@ package baseballgame;
 
 import java.util.*;
 
+import static baseballgame.Validator.VALID_NUMBER_LENGTH;
+
 public class BaseBallGame {
-	private final int VALID_NUMBER_LENGTH = 3;
+	private final Validator validator;
+
 	private String[] problemNumbers;
 	private String[] answerNumbers;
 	private int strikeCount = 0;
 	private int ballCount = 0;
 	private boolean isPlaying = false;
 
+	public BaseBallGame(Validator validator) {
+		this.validator = validator;
+	}
+
 	public void createProblemNumbers() {
 		Set<String> numberSet = createNumberSet();
 		problemNumbers = String.join("", numberSet)
-				.split("");
+							   .split("");
 
-		if (!validateStringNumbers(problemNumbers)) {
+		if (!validator.validateStringNumbers(problemNumbers)) {
 			throw new IllegalStateException("문제 생성에 실패하였습니다. 문제 생성 로직을 확인해주세요.");
 		}
 
 		isPlaying = true;
+		System.out.println(Arrays.asList(problemNumbers));
 	}
 
 	private Set<String> createNumberSet() {
 		Set<String> numbers = new LinkedHashSet<>();
 
-		while (numbers.size() < 3) {
+		while (numbers.size() < VALID_NUMBER_LENGTH) {
 			int number = drawRandomNumberFrom1to9();
 			numbers.add(String.valueOf(number));
 		}
@@ -38,25 +46,13 @@ public class BaseBallGame {
 		return random.nextInt(9) + 1;
 	}
 
-	boolean validateStringNumbers(String[] stringNumbers) {
-		if (stringNumbers == null || stringNumbers.length != 3) {
-			return false;
-		}
-
-		String NUMBER_FROM_1_to_9_REG_EXP = "^[1-9]+$";
-		String join = String.join("", stringNumbers);
-		Set<String> set = new HashSet<>(Arrays.asList(stringNumbers));
-
-		return join.matches(NUMBER_FROM_1_to_9_REG_EXP) && set.size() == VALID_NUMBER_LENGTH;
-	}
-
 	public void receiveAnswerNumbers() {
 		Scanner scanner = new Scanner(System.in);
 
 		do {
 			System.out.print("숫자를 입력해주세요(1~9 중 서로 다른 3자리수) : ");
 			answerNumbers = scanner.next().split("");
-		} while (!validateStringNumbers(answerNumbers));
+		} while (!validator.validateStringNumbers(answerNumbers));
 	}
 
 	public void countStrikesAndBalls() {
@@ -64,7 +60,7 @@ public class BaseBallGame {
 		countStrikes();
 		countBalls();
 
-		if (!validateStrikeAndBallCounts()) {
+		if (!validator.validateStrikeAndBallCounts(strikeCount, ballCount)) {
 			throw new IllegalStateException("스트라이크, 볼 판정에 오류가 발생했습니다. 로직을 확인해주세요.");
 		}
 	}
@@ -97,16 +93,6 @@ public class BaseBallGame {
 		if (!isStrike && problemNumbers.contains(answerNumber)) {
 			ballCount++;
 		}
-	}
-
-	boolean validateStrikeAndBallCounts() {
-		if ((strikeCount + ballCount) > 3 ||
-			(strikeCount == 2 && ballCount == 1) ||
-			(strikeCount == 3 && ballCount != 0)) {
-			return false;
-		}
-
-		return true;
 	}
 
 	public void printCountResults() {
@@ -148,13 +134,9 @@ public class BaseBallGame {
 		do {
 			System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
 			restartOrEndStatus = scanner.next();
-		} while (!validateRestartOrEndStatus(restartOrEndStatus));
+		} while (!validator.validateRestartOrEndStatus(restartOrEndStatus));
 
 		return restartOrEndStatus;
-	}
-
-	boolean validateRestartOrEndStatus(String restartOrEndStatus) {
-		return restartOrEndStatus.equals("1") || restartOrEndStatus.equals("2");
 	}
 
 	public boolean isPlaying() {
